Fixed errors related to weird typecasting results from google maps. Outdoor POI priceLevels and opening hours on hotspots now depend on dynamic types

+++ lib/widgets/screens/outdoor_poi_results.dart
@@ -43,7 +43,7 @@ class _OutdoorPOIResultsState extends State<OutdoorPOIResults> {
                         Row(
                           children: <Widget>[
                             Text(_results[index].name),
-                            Text(priceToString(_results[index].priceLevel)),
+                            Text(priceToString(_results[index].priceLevel as int)),
                           ],
                         ),
                         Text(_results[index].address),
